requirements.txt
diffdist==0.1
easydict==1.10
mlflow==2.4.2
numpy==1.24.3
Pillow==9.4.0
Pillow==10.1.0
scikit_learn==1.2.2
termcolor==2.4.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
